Unify the no_std and std impls of FontSystem
