+++ Cargo.toml
@@ -21,6 +21,8 @@ unicode-linebreak = "0.1.4"
 unicode-script = "0.5.5"
 unicode-segmentation = "1.10.0"
 rangemap = "1.2.0"
+hashbrown = { version = "0.14.0", default-features = false }
+rustc-hash = { version = "1.1.0", default-features = false }
 
 [dependencies.unicode-bidi]
 version = "0.3.8"
